Prod branches up to date

+++ src/pages/scoreboard.jsx
@@ -24,6 +24,12 @@ const [modalEditVisible, setEditModalVisible] = useState(false);
 const [selectedScore, setSelectedScore] = useState(null);
 
 useEffect(() => {
+  if (!isTokenValid() && token) {
+    localStorage.removeItem("token");
+    setToken(null);
+    location.reload();
+  }
+
   fetch("https://meetings-scoreboard.onrender.com/api/scoreboard")
   .then(response => {
     if(response.ok) return response.json()
@@ -38,9 +44,17 @@ useEffect(() => {
     setLoading(false);
     toast.error("Hmm capaz el server está caido, dale tiempo")
   })
+
 }, []);
 
 function modUser(data, originalSurname) {
+  if (!isTokenValid()) {
+    localStorage.removeItem("token");
+    setToken(null);
+    toast.error("Token inválido, por favor inicia sesión de nuevo");
+    return;
+  }
+
   toast.info("En eso andamos")
   fetch(`https://meetings-scoreboard.onrender.com/api/scoreboard/${encodeURIComponent(originalSurname)}`, {
     method: "PUT",
@@ -67,6 +81,13 @@ function modUser(data, originalSurname) {
 }
 
 function deleteUser(surname) {
+  if (!isTokenValid()) {
+    localStorage.removeItem("token");
+    setToken(null);
+    toast.error("Token inválido, por favor inicia sesión de nuevo");
+    return;
+  }
+
   toast.info("Eliminando a tudino, na mentira")
   fetch(`https://meetings-scoreboard.onrender.com/api/scoreboard/${surname}`, {
     method: "DELETE",
@@ -82,6 +103,13 @@ function deleteUser(surname) {
 }
 
 function createUser(data) {
+  if (!isTokenValid()) {
+    localStorage.removeItem("token");
+    setToken(null);
+    toast.error("Token inválido, por favor inicia sesión de nuevo");
+    return;
+  }
+
   toast.info("Creando usuario my friend")
   fetch("https://meetings-scoreboard.onrender.com/api/scoreboard", {
     method: "POST",
@@ -107,6 +135,21 @@ function createUser(data) {
   });
 }
 
+function isTokenValid() {
+  const token = localStorage.getItem("token");
+  if (!token) return false;
+
+  try {
+    const payload = JSON.parse(atob(token.split(".")[1]));
+    const now = Math.floor(Date.now() / 1000);
+    return payload.exp > now;
+  } catch (err) {
+    console.error("Error leyendo token:", err);
+    return false;
+  }
+
+}
+
 
   return ( 
   <section className="scoreboard">
@@ -114,7 +157,7 @@ function createUser(data) {
     {token && selectedScore && <ModModal selectedScore={selectedScore} isOpen={modalEditVisible} onClose={() => setEditModalVisible(false)} handleDelete={() => deleteUser(selectedScore.surname)} handleMod={modUser} />}
     <h1 className="scoreboard-title">TABLA DE POSICIONES</h1>
     {loading && <div className="loading">
-      <ClipLoader color="#ffffffff" size={300} />
+      <ClipLoader color="#ffffffff" size={250} />
       <h1>{texts[rand]}</h1>
       </div>}
     <div className="scoreboard-grid">
